Add MISSING_DIRECTION quirk
Some devices like VRS DirectForce Pro 20 are missing the direction
field in their descriptor. Though it's not optional, direction value
for one-axis effects ahould always be 0x4000. The only known exception
is Forza Horizon that uses direction instead of negative magnitude
values but we can live with that.

Fixes FFB for VRS DFP20